requirements.txt
flask
flask-cors
torch
torchvision
torchaudio
pillow
opencv-python
numpy